Migrated weekoverview and improved/modularised SwipeableList
