+++ utils/SharedStyles.ts
@@ -13,25 +13,21 @@ const deviceWidth = Dimensions.get("window").width;
 const deviceHeight = Dimensions.get("window").height;
 
 const ScaleSizeW = (originalSize: number) => {
-  if(deviceWidth >= 820)
-  {
-    return originalSize
+  if (deviceWidth >= 820) {
+    return originalSize;
   }
-  if(deviceWidth <= 320)
-  {
-    return originalSize / 2
+  if (deviceWidth <= 320) {
+    return originalSize / 2;
   }
   return (originalSize / 820) * deviceWidth;
 };
 
 const ScaleSizeH = (originalSize: number) => {
-  if(deviceHeight >= 1180)
-  {
-    return originalSize
+  if (deviceHeight >= 1180) {
+    return originalSize;
   }
-  if(deviceHeight <= 480)
-  {
-    return originalSize / 2
+  if (deviceHeight <= 480) {
+    return originalSize / 2;
   }
   return (originalSize / 1180) * deviceHeight;
 };
@@ -46,15 +42,13 @@ const ScaleSizeH = (originalSize: number) => {
  * ScaleSize(100) // Returns 100 scaled to the device's width
  */
 const ScaleSize = (originalSize: number) => {
-  if(deviceHeight >= 1180)
-    {
-      return originalSize
-    }
-    if(deviceHeight <= 480)
-    {
-      return originalSize / 2
-    }
-    return (originalSize / 1180) * deviceHeight;
+  if (deviceHeight >= 1180) {
+    return originalSize;
+  }
+  if (deviceHeight <= 480) {
+    return originalSize / 2;
+  }
+  return (originalSize / 1180) * deviceHeight;
 };
 
 /**
